videospider

+++ VideoSpider/spiders/xng.py
@@ -1,17 +1,20 @@
 # -*- coding: utf-8 -*-
 import base64
 import json
-import random
+import os
 import re
+import time
 from copy import deepcopy
 from pprint import pprint
+from random import choice
 
 import pymysql
 import requests
 import scrapy
-from random import choice
-from VideoSpider.settings import *
-from VideoSpider.tool import *
+
+from VideoSpider.settings import MYSQL_HOST, MYSQL_PORT, MYSQL_USERNAME, MYSQL_PASSWORK, MYSQL_DATABASE, \
+    xng_spider_dict, PROXY_URL, xng_headers
+from VideoSpider.tool import check, jieba_ping, download, download_img, oss_upload, deeimg
 
 
 class XngSpider(scrapy.Spider):
@@ -52,6 +55,14 @@ class XngSpider(scrapy.Spider):
 
             yield scrapy.Request(url, callback=self.parse, meta={'item': deepcopy(item)}, dont_filter=True)
 
+    def deep_img_video(self, width, height, y, w, h, excursion, filename, deepcopy_filename):
+        try:
+            os.system('''ffmpeg -i {} -filter_complex "delogo=x={}:y={}:w={}:h={}:show=0" {}'''.format(filename, int(
+                width) - excursion, y, w, h, deepcopy_filename))
+            return True
+        except:
+            return False
+
     def parse(self, response):
         isotimeformat = '%Y-%m-%d'
         item = response.meta['item']
@@ -74,8 +85,8 @@ class XngSpider(scrapy.Spider):
                 item['thumbnails'] = video['url']
                 item['title'] = video['title']
                 item['id'] = video['album_id']
-                item['video_height'] = video['vw']
-                item['video_width'] = video['w']
+                item['video_height'] = 0
+                item['video_width'] = 0
                 item['spider_time'] = time.strftime(isotimeformat, time.localtime(time.time()))
                 item['from'] = '小年糕'
                 item['category'] = item['category']
@@ -86,24 +97,33 @@ class XngSpider(scrapy.Spider):
                 if (result is None) and (item['view_cnt'] >= item['view_cnt_compare']) :
                     match_type = jieba_ping(item)
                     item['match_type'] = match_type
-                    filename = download(item['osskey'], item['download_url'], True)
-                    # img_filename = download_img(item['thumbnails'], item['osskey'])
-                    if filename:
+                    filename = download(item['osskey'], item['download_url'], False)
+                    img_filename = download_img(item['thumbnails'], item['osskey'])
+                    if filename and img_filename:
                         video_size = deeimg(item['download_url'])
                         if not video_size is False and(video_size[0]<video_size[1]):
-                            deeimg_filename = 'stockpile/' + item['osskey'] + '.mp4'
-                            is_ture = deep_img_video(video_size[0], video_size[1], video_size[1]-70, 100, 50, 120, filename, deeimg_filename)
+                            deeimg_filename = item['osskey'] + '.mp4'
+                            is_ture = self.deep_img_video(video_size[0], video_size[1], video_size[1]-70, 100, 50, 120, filename, deeimg_filename)
                             if is_ture is True:
                                 oss_upload(item['osskey'], deeimg_filename, video_size[2])
+                                if os.path.exists(video_size[2]):
+                                    os.remove(video_size[2])
+
+                                if os.path.exists(filename):
+                                    os.remove(filename)
+
+                                yield item
                             else:
                                 print('去水印失败')
                         else:
                             oss_upload(item['osskey'], filename, video_size[2])
+                            if os.path.exists(video_size[2]):
+                                os.remove(video_size[2])
 
-                        if os.path.exists(video_size[2]):
-                            os.remove(video_size[2])
+                            if os.path.exists(filename):
+                                os.remove(filename)
 
-                    yield item
+                            yield item
 
         except Exception as f:
             pprint(f)
